Moves minimal attribute dict to fixture
